+++ IndicatorsTests/Test.Beta.cs
@@ -25,7 +25,7 @@ namespace StockIndicators.Tests
 
             // sample value
             BetaResult r = results.Where(x => x.Index == 502).FirstOrDefault();
-            Assert.AreEqual((decimal)1.6759, Math.Round((decimal)r.Beta, 4));
+            Assert.AreEqual(1.6759m, Math.Round((decimal)r.Beta, 4));
         }
