+++ IndicatorsTests/Test.Correlation.cs
@@ -25,8 +25,8 @@ namespace StockIndicators.Tests
 
             // sample value
             CorrResult r = results.Where(x => x.Index == 502).FirstOrDefault();
-            Assert.AreEqual((decimal)0.8460, Math.Round((decimal)r.Correlation, 4));
-            Assert.AreEqual((decimal)0.7157, Math.Round((decimal)r.RSquared, 4));
+            Assert.AreEqual(0.8460m, Math.Round((decimal)r.Correlation, 4));
+            Assert.AreEqual(0.7157m, Math.Round((decimal)r.RSquared, 4));
         }
